Changed data types of variables to be more memory efficient (in engine and engine logic files).

+++ Source-Files-Folder/Chess-Engine-Folder/Header-Files-Folder/board-score-defines.h
@@ -2,27 +2,27 @@
 #ifndef BOARD_SCORE_DEFINES_H
 #define BOARD_SCORE_DEFINES_H
 
-extern const signed short PAWN_MATRIX[8][8];
+extern const int PAWN_MATRIX[8][8];
 
-extern const signed short KNIGHT_MATRIX[8][8];
+extern const int KNIGHT_MATRIX[8][8];
 
-extern const signed short BISHOP_MATRIX[8][8];
+extern const int BISHOP_MATRIX[8][8];
 
-extern const signed short ROOK_MATRIX[8][8];
+extern const int ROOK_MATRIX[8][8];
 
-extern const signed short QUEEN_MATRIX[8][8];
+extern const int QUEEN_MATRIX[8][8];
 
-extern const signed short KING_MATRIX[8][8];
+extern const int KING_MATRIX[8][8];
 
-extern const signed short PIECE_TYPE_SCORES[7];
+extern const int PIECE_TYPE_SCORES[7];
 
-extern const signed short MATE_SCORE;
-extern const signed short DRAW_SCORE;
+extern const int MATE_SCORE;
+extern const int DRAW_SCORE;
 
 extern const float MATRIX_FACTOR;
 extern const float PIECE_FACTOR;
 
-extern const signed short MIN_STATE_SCORE;
-extern const signed short MAX_STATE_SCORE;
+extern const int MIN_STATE_SCORE;
+extern const int MAX_STATE_SCORE;
 
 #endif
