+++ Source-Files-Folder/Chess-Engine-Folder/Header-Files-Folder/guess-order-moves.h
@@ -2,12 +2,12 @@
 #ifndef GUESS_ORDER_MOVES_H
 #define GUESS_ORDER_MOVES_H
 
-bool guess_moves_scores(signed short** moveScores, const Move moveArray[], short moveAmount, const Piece board[], Info info);
+bool guess_moves_scores(int** moveScores, const Move moveArray[], int moveAmount, const Piece board[], Info info);
 
-signed short guess_move_score(const Piece board[], Info info, Move move);
+int guess_move_score(const Piece board[], Info info, Move move);
 
-bool guess_order_moves(Move* moveArray, short moveAmount, const Piece board[], Info info, unsigned short team);
+bool guess_order_moves(Move* moveArray, int moveAmount, const Piece board[], Info info, uint8_t team);
 
-bool ordered_legal_moves(Move** moveArray, short* moveAmount, const Piece board[], Info info, unsigned short team);
+bool ordered_legal_moves(Move** moveArray, int* moveAmount, const Piece board[], Info info, uint8_t team);
 
 #endif
