+++ Source-Files-Folder/Chess-Engine-Folder/Header-Files-Folder/optimal-depth-move.h
@@ -4,9 +4,9 @@
 
 bool optimal_depth_move(Move* move, const Piece board[], Info info, short seconds);
 
-bool search_depths_move(Move* move, const Piece board[], Info info, int playerSign, short seconds, const Move moveArray[], short moveAmount);
+bool search_depths_move(Move* move, const Piece board[], Info info, int playerSign, short seconds, const Move moveArray[], int moveAmount);
 
-bool choose_timing_move(Move* move, signed short* score, const Piece board[], Info info, short depth, int playerSign, long startClock, short seconds, const Move moveArray[], short moveAmount);
+bool choose_timing_move(Move* move, signed short* score, const Piece board[], Info info, short depth, int playerSign, long startClock, short seconds, const Move moveArray[], int moveAmount);
 
 bool timing_limit_ended(long startClock, short seconds);
