+++ Source-Files-Folder/Engine-Logic-Folder/Header-Files-Folder/engine-info-values.h
@@ -2,7 +2,7 @@
 #ifndef ENGINE_INFO_VALUES_H
 #define ENGINE_INFO_VALUES_H
 
-typedef unsigned int Info;
+typedef uint32_t Info;
 
 extern const Info INFO_TEAM_MASK;
 extern const Info INFO_PASSANT_MASK;
@@ -19,10 +19,10 @@ extern const Info INFO_TEAM_NONE;
 extern const Info INFO_TEAM_WHITE;
 extern const Info INFO_TEAM_BLACK;
 
-extern const unsigned short INFO_TEAM_SHIFT;
-extern const unsigned short INFO_PASSANT_SHIFT;
-extern const unsigned short INFO_TURNS_SHIFT;
-extern const unsigned short INFO_COUNTER_SHIFT;
+extern const uint8_t INFO_TEAM_SHIFT;
+extern const uint8_t INFO_PASSANT_SHIFT;
+extern const uint8_t INFO_TURNS_SHIFT;
+extern const uint8_t INFO_COUNTER_SHIFT;
 
 extern const Info INFO_NONE;
 extern const Info INFO_BLANK;
