+++ Source-Files-Folder/Engine-Logic-Folder/Header-Files-Folder/engine-piece-values.h
@@ -2,7 +2,7 @@
 #ifndef ENGINE_PIECE_VALUES_H
 #define ENGINE_PIECE_VALUES_H
 
-typedef unsigned short Piece;
+typedef uint8_t Piece;
 
 extern const Piece PIECE_TYPE_NONE;
 extern const Piece PIECE_TYPE_PAWN;
@@ -19,8 +19,8 @@ extern const Piece PIECE_TEAM_BLACK;
 extern const Piece PIECE_TEAM_MASK;
 extern const Piece PIECE_TYPE_MASK;
 
-extern const unsigned short PIECE_TEAM_SHIFT;
-extern const unsigned short PIECE_TYPE_SHIFT;
+extern const uint8_t PIECE_TEAM_SHIFT;
+extern const uint8_t PIECE_TYPE_SHIFT;
 
 extern const Piece PIECE_NONE;
 extern const Piece PIECE_BLANK;
