+++ Source-Files-Folder/Engine-Logic-Folder/Header-Files-Folder/move-pattern-handler.h
@@ -4,16 +4,16 @@
 
 bool move_pattern_valid(Move move, Piece piece);
 
-bool pawn_pattern_handler(Move move, unsigned short team);
+bool pawn_pattern_handler(Move move, uint8_t team);
 
-bool knight_pattern_handler(Move move, unsigned short team);
+bool knight_pattern_handler(Move move, uint8_t team);
 
-bool bishop_pattern_handler(Move move, unsigned short team);
+bool bishop_pattern_handler(Move move, uint8_t team);
 
-bool queen_pattern_handler(Move move, unsigned short team);
+bool queen_pattern_handler(Move move, uint8_t team);
 
-bool king_pattern_handler(Move move, unsigned short team);
+bool king_pattern_handler(Move move, uint8_t team);
 
-bool rook_pattern_handler(Move move, unsigned short team);
+bool rook_pattern_handler(Move move, uint8_t team);
 
 #endif
